Input: Move generic types from joystick/ to input/ directory
