+++ xbmc/games/controllers/guicontrols/GUICardinalFeatureButton.h
@@ -38,7 +38,7 @@ namespace GAME
     // implementation of IFeatureButton
     virtual bool PromptForInput(CEvent& waitEvent) override;
     virtual bool IsFinished(void) const override;
-    virtual JOYSTICK::CARDINAL_DIRECTION GetCardinalDirection(void) const override;
+    virtual INPUT::CARDINAL_DIRECTION GetCardinalDirection(void) const override;
     virtual void Reset(void) override;
 
   private:
