+++ xbmc/games/controllers/windows/GUIConfigurationWizard.h
@@ -103,7 +103,7 @@ namespace GAME
 
     // State variables and mutex
     IFeatureButton*                      m_currentButton;
-    JOYSTICK::CARDINAL_DIRECTION         m_cardinalDirection;
+    INPUT::CARDINAL_DIRECTION            m_cardinalDirection;
     JOYSTICK::WHEEL_DIRECTION            m_wheelDirection;
     JOYSTICK::THROTTLE_DIRECTION         m_throttleDirection;
     std::set<JOYSTICK::CDriverPrimitive> m_history; // History to avoid repeated features
