+++ xbmc/input/joysticks/JoystickTranslator.cpp
@@ -27,9 +27,6 @@
 using namespace KODI;
 using namespace JOYSTICK;
 
-#define TAN_1_8_PI  0.4142136f // tan(1/8*PI)
-#define TAN_3_8_PI  2.4142136f // tan(3/8*PI)
-
 const char* CJoystickTranslator::HatStateToString(HAT_STATE state)
 {
   switch (state)
@@ -138,30 +135,6 @@ THROTTLE_DIRECTION CJoystickTranslator::PositionToThrottleDirection(float positi
   else if (position < 0.0f) return THROTTLE_DIRECTION::DOWN;
 
   return THROTTLE_DIRECTION::NONE;
-}
-
-CARDINAL_DIRECTION CJoystickTranslator::VectorToCardinalDirection(float x, float y)
-{
-  if      (y >= -x && y >  x) return CARDINAL_DIRECTION::UP;
-  else if (y <=  x && y > -x) return CARDINAL_DIRECTION::RIGHT;
-  else if (y <= -x && y <  x) return CARDINAL_DIRECTION::DOWN;
-  else if (y >=  x && y < -x) return CARDINAL_DIRECTION::LEFT;
-
-  return CARDINAL_DIRECTION::NONE;
-}
-
-INTERCARDINAL_DIRECTION CJoystickTranslator::VectorToIntercardinalDirection(float x, float y)
-{
-  if      (y >= TAN_3_8_PI * -x && y > TAN_3_8_PI *  x) return INTERCARDINAL_DIRECTION::UP;
-  else if (y <= TAN_3_8_PI *  x && y > TAN_1_8_PI *  x) return INTERCARDINAL_DIRECTION::RIGHTUP;
-  else if (y <= TAN_1_8_PI *  x && y > TAN_1_8_PI * -x) return INTERCARDINAL_DIRECTION::RIGHT;
-  else if (y <= TAN_1_8_PI * -x && y > TAN_3_8_PI * -x) return INTERCARDINAL_DIRECTION::RIGHTDOWN;
-  else if (y <= TAN_3_8_PI * -x && y < TAN_3_8_PI *  x) return INTERCARDINAL_DIRECTION::DOWN;
-  else if (y >= TAN_3_8_PI *  x && y < TAN_1_8_PI *  x) return INTERCARDINAL_DIRECTION::LEFTDOWN;
-  else if (y >= TAN_1_8_PI *  x && y < TAN_1_8_PI * -x) return INTERCARDINAL_DIRECTION::LEFT;
-  else if (y >= TAN_1_8_PI * -x && y < TAN_3_8_PI * -x) return INTERCARDINAL_DIRECTION::LEFTUP;
-
-  return INTERCARDINAL_DIRECTION::NONE;
 }
 
 std::string CJoystickTranslator::GetPrimitiveName(const CDriverPrimitive& primitive)
