+++ xbmc/input/joysticks/JoystickTranslator.h
@@ -125,34 +125,6 @@ namespace JOYSTICK
     static THROTTLE_DIRECTION PositionToThrottleDirection(float position);
 
     /*!
-     * \brief Get the closest cardinal direction to the given vector
-     *
-     * Ties are resolved in the clockwise direction: (0.5, 0.5) will resolve to
-     * RIGHT.
-     *
-     * \param x  The x component of the vector
-     * \param y  The y component of the vector
-     *
-     * \return The closest cardinal direction (up, down, right or left), or
-     *         CARDINAL_DIRECTION::UNKNOWN if x and y are both 0
-     */
-    static CARDINAL_DIRECTION VectorToCardinalDirection(float x, float y);
-
-    /*!
-     * \brief Get the closest cardinal or intercardinal direction to the given
-     *        vector
-     *
-     * Ties are resolved in the clockwise direction.
-     *
-     * \param x  The x component of the vector
-     * \param y  The y component of the vector
-     *
-     * \return The closest intercardinal direction, or
-     *         INTERCARDINAL_DIRECTION::UNPRESSED if x and y are both 0
-     */
-    static INTERCARDINAL_DIRECTION VectorToIntercardinalDirection(float x, float y);
-
-    /*!
      * \brief Get the localized name of the primitive
      *
      * \param primitive The primitive, currently only buttons and axes are supported
